refactor: restructure streaming implementation
- Extract streaming logic into separate files
- Add base structure for manual tool control
- Separate common onFinish handler
- Prepare for non-tool-calling models support
