+++ app/api/chat/route.ts
@@ -1,15 +1,6 @@
-import { getChat, saveChat } from '@/lib/actions/chat'
-import { generateRelatedQuestions } from '@/lib/agents/generate-related-questions'
-import { researcher } from '@/lib/agents/researcher'
-import { ExtendedCoreMessage } from '@/lib/types'
-import { convertToExtendedCoreMessages } from '@/lib/utils'
+import { createManualToolStreamResponse } from '@/lib/streaming/create-manual-tool-stream'
+import { createToolCallingStreamResponse } from '@/lib/streaming/create-tool-calling-stream'
 import { isProviderEnabled } from '@/lib/utils/registry'
-import {
-  convertToCoreMessages,
-  createDataStreamResponse,
-  JSONValue,
-  streamText
-} from 'ai'
 import { cookies } from 'next/headers'
 
 export const maxDuration = 30
@@ -29,9 +20,6 @@ export async function POST(req: Request) {
       })
     }
 
-    const coreMessages = convertToCoreMessages(messages)
-    const extendedCoreMessages = convertToExtendedCoreMessages(messages)
-
     const cookieStore = await cookies()
     const modelFromCookie = cookieStore.get('selected-model')?.value
     const model = modelFromCookie || DEFAULT_MODEL
@@ -44,97 +32,19 @@ export async function POST(req: Request) {
       })
     }
 
-    return createDataStreamResponse({
-      execute: async dataStream => {
-        try {
-          let researcherConfig
-          try {
-            researcherConfig = await researcher({
-              messages: coreMessages,
-              model
-            })
-          } catch (error) {
-            console.error('Researcher configuration error:', error)
-            throw new Error('Failed to initialize researcher configuration')
-          }
+    const supportsToolCalling = provider === 'openai'
 
-          const result = streamText({
-            ...researcherConfig,
-            onFinish: async event => {
-              try {
-                const responseMessages = event.response.messages
-
-                let annotation: JSONValue = {
-                  type: 'related-questions',
-                  data: {
-                    items: []
-                  }
-                }
-
-                // Notify related questions loading
-                dataStream.writeMessageAnnotation(annotation)
-
-                // Generate related questions
-                const relatedQuestions = await generateRelatedQuestions(
-                  responseMessages,
-                  model
-                )
-
-                // Update the annotation with the related questions
-                annotation = {
-                  ...annotation,
-                  data: relatedQuestions.object
-                }
-
-                // Send related questions to client
-                dataStream.writeMessageAnnotation(annotation)
-
-                // Create the message to save
-                const generatedMessages = [
-                  ...extendedCoreMessages,
-                  ...responseMessages.slice(0, -1),
-                  {
-                    role: 'data',
-                    content: annotation
-                  },
-                  responseMessages[responseMessages.length - 1]
-                ] as ExtendedCoreMessage[]
-
-                // Get the chat from the database if it exists, otherwise create a new one
-                const savedChat = (await getChat(chatId)) ?? {
-                  messages: [],
-                  createdAt: new Date(),
-                  userId: 'anonymous',
-                  path: `/search/${chatId}`,
-                  title: messages[0].content,
-                  id: chatId
-                }
-
-                // Save chat with complete response and related questions
-                await saveChat({
-                  ...savedChat,
-                  messages: generatedMessages
-                }).catch(error => {
-                  console.error('Failed to save chat:', error)
-                  throw new Error('Failed to save chat history')
-                })
-              } catch (error) {
-                console.error('Error in onFinish:', error)
-                throw error
-              }
-            }
-          })
-
-          result.mergeIntoDataStream(dataStream)
-        } catch (error) {
-          console.error('Stream execution error:', error)
-        }
-      },
-      onError: error => {
-        console.error('Stream error:', error)
-        return error instanceof Error ? error.message : String(error)
-      }
-    })
+    return supportsToolCalling
+      ? createToolCallingStreamResponse({
+          messages,
+          model,
+          chatId
+        })
+      : createManualToolStreamResponse({
+          messages,
+          model,
+          chatId
+        })
   } catch (error) {
     console.error('API route error:', error)
     return new Response(
