+++ lib/streaming/create-manual-tool-stream.ts
@@ -1,0 +1,18 @@
+import { createDataStreamResponse, DataStreamWriter } from 'ai'
+import { BaseStreamConfig } from './types'
+
+export function createManualToolStreamResponse(config: BaseStreamConfig) {
+  return createDataStreamResponse({
+    execute: async (dataStream: DataStreamWriter) => {
+      try {
+        // TODO: Implement manual tool calling
+      } catch (error) {
+        console.error('Stream execution error:', error)
+      }
+    },
+    onError: error => {
+      console.error('Stream error:', error)
+      return error instanceof Error ? error.message : String(error)
+    }
+  })
+}
