+++ lib/streaming/create-tool-calling-stream.ts
@@ -1,0 +1,46 @@
+import { researcher } from '@/lib/agents/researcher'
+import {
+  convertToCoreMessages,
+  createDataStreamResponse,
+  DataStreamWriter,
+  streamText
+} from 'ai'
+import { handleStreamFinish } from './handle-stream-finish'
+import { BaseStreamConfig } from './types'
+
+export function createToolCallingStreamResponse(config: BaseStreamConfig) {
+  return createDataStreamResponse({
+    execute: async (dataStream: DataStreamWriter) => {
+      try {
+        const coreMessages = convertToCoreMessages(config.messages)
+
+        let researcherConfig = await researcher({
+          messages: coreMessages,
+          model: config.model
+        })
+
+        const result = streamText({
+          ...researcherConfig,
+          onFinish: async result => {
+            await handleStreamFinish({
+              responseMessages: result.response.messages,
+              originalMessages: config.messages,
+              model: config.model,
+              chatId: config.chatId,
+              dataStream
+            })
+          }
+        })
+
+        result.mergeIntoDataStream(dataStream)
+      } catch (error) {
+        console.error('Stream execution error:', error)
+        throw error
+      }
+    },
+    onError: error => {
+      console.error('Stream error:', error)
+      return error instanceof Error ? error.message : String(error)
+    }
+  })
+}
