+++ lib/streaming/handle-stream-finish.ts
@@ -1,0 +1,83 @@
+import { getChat, saveChat } from '@/lib/actions/chat'
+import { generateRelatedQuestions } from '@/lib/agents/generate-related-questions'
+import { ExtendedCoreMessage } from '@/lib/types'
+import { convertToExtendedCoreMessages } from '@/lib/utils'
+import { CoreMessage, DataStreamWriter, JSONValue, Message } from 'ai'
+
+interface HandleStreamFinishParams {
+  responseMessages: CoreMessage[]
+  originalMessages: Message[]
+  model: string
+  chatId: string
+  dataStream: DataStreamWriter
+}
+
+export async function handleStreamFinish({
+  responseMessages,
+  originalMessages,
+  model,
+  chatId,
+  dataStream
+}: HandleStreamFinishParams) {
+  try {
+    const extendedCoreMessages = convertToExtendedCoreMessages(originalMessages)
+
+    let annotation: JSONValue = {
+      type: 'related-questions',
+      data: {
+        items: []
+      }
+    }
+
+    // Notify related questions loading
+    dataStream.writeMessageAnnotation(annotation)
+
+    // Generate related questions
+    const relatedQuestions = await generateRelatedQuestions(
+      responseMessages,
+      model
+    )
+
+    // Update the annotation with the related questions
+    annotation = {
+      ...annotation,
+      data: relatedQuestions.object
+    }
+
+    // Send related questions to client
+    dataStream.writeMessageAnnotation(annotation)
+
+    // Create the message to save
+    const generatedMessages = [
+      ...extendedCoreMessages,
+      ...responseMessages.slice(0, -1),
+      {
+        role: 'data',
+        content: annotation
+      } as ExtendedCoreMessage,
+      responseMessages[responseMessages.length - 1]
+    ] as ExtendedCoreMessage[]
+
+    // Get the chat from the database if it exists, otherwise create a new one
+    const savedChat = (await getChat(chatId)) ?? {
+      messages: [],
+      createdAt: new Date(),
+      userId: 'anonymous',
+      path: `/search/${chatId}`,
+      title: originalMessages[0].content,
+      id: chatId
+    }
+
+    // Save chat with complete response and related questions
+    await saveChat({
+      ...savedChat,
+      messages: generatedMessages
+    }).catch(error => {
+      console.error('Failed to save chat:', error)
+      throw new Error('Failed to save chat history')
+    })
+  } catch (error) {
+    console.error('Error in handleStreamFinish:', error)
+    throw error
+  }
+}
