+++ lib/streaming/types.ts
@@ -1,0 +1,7 @@
+import { Message } from 'ai'
+
+export interface BaseStreamConfig {
+  messages: Message[]
+  model: string
+  chatId: string
+}
